Carts Done

+++ app/src/main/java/com/android/sivano/domin/model/GetAllCarts.kt
@@ -4,6 +4,6 @@ import com.android.sivano.entities.CartItemDto
 
 data class GetAllCarts(
     val cart_items: List<CartItem>,
-    val sub_total: Int,
-    val total: Int
+    val sub_total: Double,
+    val total: Double
 )
