+++ app/src/main/java/com/android/sivano/entities/GetCartsDto.kt
@@ -2,6 +2,6 @@ package com.android.sivano.entities
 
 data class GetCartsDto(
     val cart_items: List<CartItemDto>,
-    val sub_total: Int,
-    val total: Int
+    val sub_total: Double,
+    val total: Double
 )
